Use `usize` instead of `u32` for `Int`s.
This may not be a universally good idea because it risks coupling the
semantics of the language being compiled with the host language of the
compiler, but it does simplify things and this is a toy example, so...

+++ src/lex.rs
@@ -5,7 +5,7 @@ pub(crate) enum Token<'b> {
     BSlash,        // "\"
     Comma,         // ","
     Dot,           // "."
-    Int(u32),      // integer literals
+    Int(usize),    // integer literals
     Word(&'b str), // identifiers
     Bra,           // "["
     Ket,           // "]"
@@ -68,7 +68,7 @@ impl<'b> Iterator for Lexer<'b> {
             ')' => Some(Token::RPar),
 
             n if n.is_numeric() => {
-                let int: u32 = self.gather(ix, |c| c.is_numeric()).parse().unwrap();
+                let int: usize = self.gather(ix, |c| c.is_numeric()).parse().unwrap();
                 Some(Token::Int(int))
             }
